Thunks: do not sign-extend the guest signature pointer on i386
RegisterCallbackUnpacker marshals its arguments through a struct of two
uint64_t fields so that 32-bit and 64-bit guests produce an identical
byte layout. The signature pointer was widened with a direct
  reinterpret_cast<uint64_t>(signature_name)
which on i386 sign-extends. Guest libraries are mapped above 2GiB, so a
real pointer such as 0xf777d020 reached the host as 0xfffffffff777d020
and the host faulted as soon as it dereferenced it:

  #0 strlen
  #1 std::char_traits<char>::length (__s=0xfffffffff777d020)
  #2 basic_string(...)
  #3 FEX::HLE::ThunkFunctions::RegisterCallbackUnpacker

The host cannot defend against this by masking, because a 64-bit guest
legitimately passes pointers wider than 32 bits. Widen from uintptr_t
instead: it is unsigned and exactly pointer-width on both guests, so the
conversion is a zero-extension.

This path had never executed. 32-bit guests silently failed to thunk at
all on an Arch rootfs until the /usr/lib32 overlay prefixes were added,
so the first 32-bit guest to reach the callback registry crashed on its
first registration.

+++ ThunkLibs/include/common/Guest.h
@@ -237,7 +237,14 @@ inline void RegisterCallbackUnpacker(const char* signature_name, uintptr_t guest
     uint64_t signature_name;
     uint64_t guest_unpacker;
   };
-  args_t args = {reinterpret_cast<uint64_t>(signature_name), guest_unpacker};
+  // Go through uintptr_t rather than casting the pointer straight to uint64_t.
+  // On i386 a direct reinterpret_cast<uint64_t> of a pointer sign-extends, so
+  // any guest address with the high bit set - which is where the guest's
+  // libraries actually live, e.g. 0xf777d020 - arrives at the host as
+  // 0xfffffffff777d020 and faults the moment the host dereferences it.
+  // uintptr_t is unsigned and exactly pointer-width on both guests, so widening
+  // from it is a plain zero-extension.
+  args_t args = {static_cast<uint64_t>(reinterpret_cast<uintptr_t>(signature_name)), static_cast<uint64_t>(guest_unpacker)};
   fexthunks_fex_register_callback_unpacker(&args);
 }
